refactor: move structs to utils
