+++ src/ape_console/_cli.py
@@ -13,7 +13,7 @@ import IPython  # type: ignore
 from ape import config
 from ape import project as default_project
 from ape.cli import NetworkBoundCommand, ape_cli_context, network_option
-from ape.utils import _python_version
+from ape.utils.misc import _python_version
 from ape.version import version as ape_version  # type: ignore
 
 CONSOLE_EXTRAS_FILENAME = "ape_console_extras.py"
